Migrate Redis to indy plugin
