+++ apm-agent-plugins/apm-redis-plugin/apm-jedis-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.redis.jedis.JedisInstrumentation
-co.elastic.apm.agent.redis.jedis.JedisSpanNameInstrumentation
+co.elastic.apm.agent.jedis.JedisInstrumentation
+co.elastic.apm.agent.jedis.JedisSpanNameInstrumentation
