+++ apm-agent-plugins/apm-redis-plugin/apm-lettuce-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,8 +1,8 @@
-co.elastic.apm.agent.redis.lettuce.Lettuce34StartSpanInstrumentation
-co.elastic.apm.agent.redis.lettuce.Lettuce34StopSpanInstrumentation$OnComplete
-co.elastic.apm.agent.redis.lettuce.Lettuce34StopSpanInstrumentation$OnCompleteExceptionally
-co.elastic.apm.agent.redis.lettuce.Lettuce34StopSpanInstrumentation$OnCancel
-co.elastic.apm.agent.redis.lettuce.Lettuce5StartSpanInstrumentation
-co.elastic.apm.agent.redis.lettuce.Lettuce5StopSpanInstrumentation$OnComplete
-co.elastic.apm.agent.redis.lettuce.Lettuce5StopSpanInstrumentation$OnCompleteExceptionally
-co.elastic.apm.agent.redis.lettuce.Lettuce5StopSpanInstrumentation$OnCancel
+co.elastic.apm.agent.lettuce.Lettuce34StartSpanInstrumentation
+co.elastic.apm.agent.lettuce.Lettuce34StopSpanInstrumentation$OnComplete
+co.elastic.apm.agent.lettuce.Lettuce34StopSpanInstrumentation$OnCompleteExceptionally
+co.elastic.apm.agent.lettuce.Lettuce34StopSpanInstrumentation$OnCancel
+co.elastic.apm.agent.lettuce.Lettuce5StartSpanInstrumentation
+co.elastic.apm.agent.lettuce.Lettuce5StopSpanInstrumentation$OnComplete
+co.elastic.apm.agent.lettuce.Lettuce5StopSpanInstrumentation$OnCompleteExceptionally
+co.elastic.apm.agent.lettuce.Lettuce5StopSpanInstrumentation$OnCancel
