+++ apm-agent-plugins/apm-redis-plugin/apm-redisson-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,1 +1,1 @@
-co.elastic.apm.agent.redis.redisson.RedisConnectionInstrumentation
+co.elastic.apm.agent.redisson.RedisConnectionInstrumentation
